refactor: move config.pages logic to addProgrammaticPages.ts
Move addProgrammaticPages() (and its slug/locationId helpers) out of
getPlusFilesByLocationId.ts into a standalone file. getPlusFilesByLocationId.ts
now exports getPlusFileFromConfigFile() and the PlusFileConfig type for reuse.
No behavior change.
